# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Practice MERN API
  version: 1.0.0
  description: API for user registration, login, update, and delete

servers:
  - url: http://127.0.0.1:3001

paths:
  /register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User registered or already exists

  /login:
    post:
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login success or failure

  /update:
    put:
      summary: Update user name or password (requires login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                name:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: User info updated
        '401':
          description: Unauthorized (not logged in)

  /delete:
    delete:
      summary: Delete a user by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
      responses:
        '200':
          description: User deleted